refactor(neon_notes): Refactor package structure
