+++ packages/neon/neon_notes/lib/src/blocs/notes.dart
@@ -1,4 +1,12 @@
-part of '../neon_notes.dart';
+import 'dart:async';
+
+import 'package:built_collection/built_collection.dart';
+import 'package:neon/blocs.dart';
+import 'package:neon/models.dart';
+import 'package:neon/utils.dart';
+import 'package:neon_notes/src/options.dart';
+import 'package:nextcloud/notes.dart' as notes;
+import 'package:rxdart/rxdart.dart';
 
 abstract interface class NotesBlocEvents {
   void createNote({
