+++ packages/neon/neon_notes/lib/src/utils/category_color.dart
@@ -1,4 +1,8 @@
-part of '../neon_notes.dart';
+import 'dart:convert';
+
+import 'package:crypto/crypto.dart';
+import 'package:flutter/painting.dart';
+import 'package:neon/utils.dart';
 
 class NotesCategoryColor {
   const NotesCategoryColor._();
