+++ packages/neon/neon_notes/lib/src/utils/exception_handler.dart
@@ -1,4 +1,7 @@
-part of '../neon_notes.dart';
+import 'package:flutter/material.dart';
+import 'package:neon/widgets.dart';
+import 'package:neon_notes/l10n/localizations.dart';
+import 'package:nextcloud/nextcloud.dart';
 
 void handleNotesException(final BuildContext context, final Object error) {
   if (error is DynamiteApiException && error.statusCode == 412) {
