+++ packages/neon/neon_notes/lib/src/widgets/category_select.dart
@@ -1,4 +1,7 @@
-part of '../neon_notes.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_material_design_icons/flutter_material_design_icons.dart';
+import 'package:neon_notes/l10n/localizations.dart';
+import 'package:neon_notes/src/utils/category_color.dart';
 
 class NotesCategorySelect extends StatelessWidget {
   NotesCategorySelect({
